Rename: 컴포넌트 디렉토리 이동
