+++ containers/board/SearchBar/index.module.scss
